privnet/autoexteps: Add vsphere provider
This adds a vsphere datasource provider to the auto-external-eps
subsystem. This provider periodically polls a user-configured folder in
the vSphere API and emits external endpoints for all VMs which have
known IPs and match the user-provided properties filter.

The provider is not registered in the Hive because the govomomi
dependency adds 11MB to the operator build. We still merge it into the
tree to avoid bit-rot, but it will not be used outside of unit tests.

Building on top of the private-networks statedb table introduced in the
previous commits, let's extend the logic to remap the network name,
based on the configured mappings. The creation of a given external
endpoint is skipped, emitting a warning, In case a network cannot be
found, or multiple networks match the port group name.

Co-authored-by: Marco Iorio <marco.iorio@isovalent.com>
Signed-off-by: Sebastian Wicki <sebastian@isovalent.com>