privnet/autoexteps: Add helm scaffholding
This adds the minimum set of changes required to enable the automatic
creation of external endpoints. To configure this feature, the user is
expected to provide a custom config via Helm and a set of secrets with
credentials, for example as follows:

```yaml
apiVersion: v1
kind: Secret
metadata:
  name: vsphere-creds
  namespace: cilium-secrets
type: Opaque
data:
  url: <url>
  user: <user>
  password: <password>
  cacert: <cacert>
  insecureSkipVerify: <true/false>
```

To enable automatic creation of external endpoints using the above
settings, Helm values should be set as follows:

```yaml
enterprise:
  privateNetworks:
    enabled: true
    autoExternalEndpoints:
      enabled: true
      providers:
        - type: vsphere
          name: my-vsphere
          config:
            credentials-secret-name: gvcsim-creds
            data-center: DC0
            root-folder: /DC0/vm/
            property-filter:
              - name: privnet-vm-*

```

The feature is currently left undocumented on purpose in the Helm
documentation, as there is no provider that can actaully be enabled
outside of unit tests.

Signed-off-by: Sebastian Wicki <sebastian@isovalent.com>