privnet/operator: store portgroup mappings in private-networks table
Extend the private-networks statedb table to additionally store the
mappings between each private network name and the corresponding
identifier in the target providers (only vSphere is supported at
the moment).

Note that, while this logic made sense when we had a 1:1 mapping between
a private network and a portgroup, it falls short now after the privnet
as VPC lift, and is likely to require substantial refactoring. At the
very minimum, it should be possible to associate multiple porgroups to
a single private network, either listing them all in the annotation, or
leveraging tags or other grouping properties.

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>